Implemented some static casts to narrow down ints to a float

+++ src/WalkAndRunAnimations.h
@@ -4,8 +4,8 @@
 #include <ctime>
 #include "Debug.h"
 
-int frameWidth;
-int frameHeight;
+float frameWidth;
+float frameHeight;
 Rectangle sourceRec;
 bool isRunning = false;
 int currentFrame = 0;
